+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- **CUMULUS-4898**
+  - Add Iceberg API documentation page to Cumulus Documentation
 - **CUMULUS-4866**
   - Add cmrProvider to sns output granules, executions, pdrs
 - **CUMULUS-4883**